test(#5015): pin restart retry identity reuse
